requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
notebook>=6.4.0
jupyter>=1.0.0
google-api-python-client>=2.40.0
python-dotenv>=0.19.0
isodate>=0.6.0
streamlit>=1.20.0
joblib>=1.1.0
xgboost>=1.6.0
catboost>=1.1.0
lightgbm>=3.3.0
scipy>=1.7.0
plotly>=5.0.0
